fix: guard null PR number in audit_headless.sh and td_cli.py
- Update audit_headless.sh to use jq // empty and add a bash loop guard to skip null/empty PR numbers with a warning to stderr.
- Enhance handle_audit_pr in td_cli.py with a defensive guard that raises CLIError for null, empty, 'None', or non-numeric PR numbers.
- Add regression tests in tests/dev-tools/test_td_cli.py.

Fixes ValueError: invalid literal for int() with base 10: 'null' in Mass Audit PRs workflow.
