+++ tests/dev-tools/test_td_cli.py
@@ -92,3 +92,48 @@ class TestTDCLI(unittest.TestCase):
 
 if __name__ == '__main__':
     unittest.main()
+
+class TestTDCliCrash(unittest.TestCase):
+    @patch('td_cli.get_github_client')
+    @patch('td_cli.get_repo_name')
+    def test_handle_audit_pr_null_error(self, mock_repo, mock_client):
+        args = MagicMock()
+        args.pr_number = "null"
+        args.fetch = True
+
+        with self.assertRaises(td_cli.CLIError) as cm:
+            td_cli.handle_audit_pr(args)
+        self.assertIn("Invalid PR number", cm.exception.message)
+
+    @patch('td_cli.get_github_client')
+    @patch('td_cli.get_repo_name')
+    def test_handle_audit_pr_none_error(self, mock_repo, mock_client):
+        args = MagicMock()
+        args.pr_number = None
+        args.fetch = True
+
+        with self.assertRaises(td_cli.CLIError) as cm:
+            td_cli.handle_audit_pr(args)
+        self.assertIn("Invalid PR number", cm.exception.message)
+
+    @patch('td_cli.get_github_client')
+    @patch('td_cli.get_repo_name')
+    def test_handle_audit_pr_empty_error(self, mock_repo, mock_client):
+        args = MagicMock()
+        args.pr_number = ""
+        args.fetch = True
+
+        with self.assertRaises(td_cli.CLIError) as cm:
+            td_cli.handle_audit_pr(args)
+        self.assertIn("Invalid PR number", cm.exception.message)
+
+    @patch('td_cli.get_github_client')
+    @patch('td_cli.get_repo_name')
+    def test_handle_audit_pr_non_numeric_error(self, mock_repo, mock_client):
+        args = MagicMock()
+        args.pr_number = "abc"
+        args.fetch = True
+
+        with self.assertRaises(td_cli.CLIError) as cm:
+            td_cli.handle_audit_pr(args)
+        self.assertIn("Invalid PR number format", cm.exception.message)
